client/requirements.txt
altgraph==0.16.1
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
psutil==5.6.7
PyInstaller==3.5
pylint==2.4.4
python-crontab==2.4.0
python-dateutil==2.8.1
six==1.13.0
typed-ast==1.4.0
wrapt==1.11.2
